move validation to validation folder
