+++ apps/asap-server/src/handlers/webhooks/webhook-connect-by-code.ts
@@ -1,31 +1,9 @@
 import { SquidexGraphql } from '@asap-hub/squidex';
 import { framework as lambda } from '@asap-hub/services-common';
-
-import { JSONSchemaType } from 'ajv';
 import Users from '../../controllers/users';
 import { Handler } from '../../utils/types';
 import validateRequest from '../../utils/validate-auth0-request';
-import { validateInput } from '../../validation';
-
-type Body = {
-  code: string;
-  userId: string;
-};
-
-const bodySchema: JSONSchemaType<Body> = {
-  type: 'object',
-  properties: {
-    code: { type: 'string' },
-    userId: { type: 'string' },
-  },
-  required: ['code', 'userId'],
-  additionalProperties: false,
-};
-
-const validateBody = validateInput(bodySchema, {
-  skipNull: false,
-  coerce: true,
-});
+import { validateBody } from '../../validation/webhook-connect-by-code.validation';
 
 export const handler: Handler = lambda.http(async (request) => {
   await validateRequest(request);
