+++ apps/asap-server/src/validation/webhook-connect-by-code.validation.ts
@@ -1,0 +1,20 @@
+import { JSONSchemaType } from 'ajv';
+import { validateInput } from './index';
+
+type Body = {
+  code: string;
+  userId: string;
+};
+const bodySchema: JSONSchemaType<Body> = {
+  type: 'object',
+  properties: {
+    code: { type: 'string' },
+    userId: { type: 'string' },
+  },
+  required: ['code', 'userId'],
+  additionalProperties: false,
+};
+export const validateBody = validateInput(bodySchema, {
+  skipNull: false,
+  coerce: true,
+});
